[FIX] account_3way_match: bill to validate journal dashboard
When having account_3way_match installed, the number and sum of
bill to validate is not consistent with the bills displayed when
clicking on the link.
With this commit, we adapt the query to be consistent with
`view_account_invoice_filter` which has been override.

Steps:

- Create 3 draft bills, $200 each:
	1. invoice_date_due < today, release_to_pay != 'yes
	2. invoice_date_due >= today, release_to_pay == 'yes'
	3. invoice_date_due >= today, release_to_pay != 'yes'
- Go to journal dashboard
-> On the vendor bill journal, we display 3 bills to validate and $600,
   although when clicking on the link we display only bills 1 and 2

opw-4182523

closes odoo/enterprise#70388

Related: odoo/odoo#182182
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>