[IMP] {sale_,}stock: add extra product in returns
Possibility to make a return from a picking before the validation. If
the picking is linked to a sales order, adding new product in the return
will add them into the sales order with a negative quantity. This will
be mainly used in food and beverage business to handle reusable packaging.

closes odoo/enterprise#70396

Task: 4191982
Related: odoo/odoo#180633
Related: odoo/upgrade#6549
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>