[REF] knowledge: refactor comments
Use the new tools provided by html_editor (the new editor) to handle comments in
Knowledge.

Comments anchors:

Previously, a comment was anchored around HTML content through `<span>` nodes
generated around text nodes in a selection. It was a pretty heavy operation, and
the length of the html string to store was heavily impacted by each comment.
With the new system, comments are anchored between 2 "beacons", which are 2
`<a>` self-closing nodes. The first one is placed at the anchor position of the
selection, and the second one is placed at the focus position. As long as there
is pertinent content between the 2 beacons, they persist, even through edition.
Once the content between them becomes empty, they are removed. This is much
lighter in terms of volume to save in the database.

Comments service:

A Comments service is introduced, which manipulates a reactive object containing
all data related to comments for the multiple knowledge components which use it.
It centralizes all server communications related to comments in batches, which
significantly reduces the amount of rpc's sent to the server when loading
threads and messages.

Comments plugin:

html_editor plugin, which handles the creation of beacons, and the editor
interactions around them.

Overlays:

Both the HtmlViewer and the CommentsPlugin use overlays to display comments to
the user, both in the HtmlField value (text is highlighted), and around it
through floating "Comment boxes", which are moving accordingly with the user
interactions.

Reused existing concepts:

Comment Handler stays, it still displays "Comment boxes" related to beacons
present in the HtmlField value.

Comment Panel stays, it still displays all comments related to the current
article

The comment panel now has a "load more" feature, which means the server won't be
returning all threads since the origins of time when an article is loaded.

task-3672747

Part-of: odoo/enterprise#67083
Related: odoo/odoo#173997
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>