[REF] knowledge: update the action helper and the template picker dialog
This commit uses the new editor features and embedded components for the action
helper and the template picker.

The template picker now uses the HtmlViewer to display prefabricated articles
with embedded components, which can be used as the default value for an article
and automatically create related article items.

The action helper provides a menu to help the user start using Knowledge.
Features such as: inserting a template, an embedded view for article items,
using ChatGPT to write content.

task-3672747

Part-of: odoo/enterprise#67083
Related: odoo/odoo#173997
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>