[IMP] sale_subscription: closing end of period
Before this PR, subscription plans lacked an option to specify whether the
subscription should close immediately or at the end of the subscription period.

After this PR, a new field has been added to subscription plans, allowing
users to choose whether a subscription should close immediately or at the end of
it's period.

task-4035347

closes odoo/enterprise#68429

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>