[IMP] appointment: use naive UTC for reference date
We now display custom appointments in the kanban view, which requires comparing
the reference date with the slot's end time excluding timezone information to
calculate _gantt_unavailability.

This commit ensures that when _slots_generate is called from _gantt_unavailability,
it receives a naive UTC datetime for the reference date.

Task-3919395

Part-of: odoo/enterprise#69042
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>