[FIX] account_accountant: no sequence on draft deferred entries
When automatic entries are generated, a sequence number should not be
assigned on them as long as they are on draft state. This is done because
it may be confusing to see a sequence name for a move in draft.
This commit ensures that the name of the entries is 'Draft' (or '/')
when the move is generated. A sequence number will be assigned once the
move is posted. Before this commit, it was possible to generate a
draft move with a sequence number when doing a deferred entries with
some entries in the future.

task-4069862

closes odoo/enterprise#70855

Related: odoo/odoo#181779
Signed-off-by: John Laterre (jol) <jol@odoo.com>