[FIX] account_reports: prevent modification of dict during iteration
The `RuntimeError: dictionary changed size during iteration` occurred in
the `_generate_tax_closing_entries` method when iterating over the
`closing_moves_by_company` dictionary. This happened because accessing a
key in a `defaultdict` that doesn't exist automatically creates that key
with a default value. Specifically, the line:
```py
if from_post and move == closing_moves_by_company[self.env.company]:
```
triggered the insertion of a new key when `self.env.company` was not
already present, leading to the runtime error.

This commit changes the access method to use `.get()` instead of
directly accessing the key allowing, safe retrieval of the value without
modifying the dictionary.

opw-4089738

closes odoo/enterprise#70944

X-original-commit: 3f5ff9767353444c476c16371273a3b3982a9d67
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>