[FIX] web_studio: discriminate approvers from delegation
Before this commit, when an admin put some users as approvers in the Rule's form
view, and then delegated, all the approvers were affected. If the delegation had a date,
then, at the end of that date, no approvers would be valid on the rule.

After this commit, with the help of a new field, we know that an approver comes from a delegation, and
we can leave the rest as it was.

task-4203697

Part-of: odoo/enterprise#70511
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>