[FIX] project_sale_subscription : Fetch profit data even if using different analytic plan than default
Steps to reproduce:
	- Create a new project and set it to be billable.
	- Set an analytic account on it
	- On the analytic account, set a plan that is not the default plan
	- Create a sale order for a subscription product with this analytic account and confirm it.
	- Create an invoice and confirm it.
	- Go to the project and click on Status Updates
	- Notice that the revenues show 0 in “Invoiced” even though we have invoiced the product
	- If we change the plan on the analytic account, we will be able to see the invoiced amount on the profitability
	report

Current behavior before PR:
When setting analytic account with different analytic plan than the
default one the analytic account value will be stored in x_plan{plan_id}_id
field when creating the analytic account line record.
https://github.com/odoo/odoo/blob/39029710bbce55889c6b951fc423c1254e05ff22/addons/account/models/account_move_line.py#L3092

https://github.com/odoo/odoo/blob/17.0/addons/analytic/models/analytic_plan.py#L106:L122

So when fetching the profitability data for a project we are just checking
the account_id field.
https://github.com/odoo/enterprise/blob/17.0/project_sale_subscription/models/project.py#L111:L115

Desired behavior after PR is merged:
We are now using the computed field auto_account_id to know which field should we
use in the domain whether it is 'account_id' or 'x_plan{plan_id}_id'

opw-4137931

closes odoo/enterprise#70508

Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>