[FIX] industry_fsm: make sure to go to backend with the right action
Before this commit, when the user goes to fsm task, start a timer and
stop it and then sign report. The user will go to the portal view, if he
clicks on the button to go back to the backend view, he first no longer
have tags, SO, project, ... fields hidden if he was in mobile view.
Second, he has no longer the different menus available in the mobile dashboard
or the menu bar in desktop.

This commit makes sure to load the main action to correctly get the
different menu in FSM when the user goes back to the backend view.

task-4189791

closes odoo/enterprise#70353

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>