[REF] sign: remove user_context usage from kanban
Keys of the user_context are available in the context, so having
the user_context available in kanban archs is redundant. There was
only 2 usecases, both in sign. This commit removes them such that
we can simplify the rendering context of kanban cards by removing
the user_context from it.

Part of task~3992107

closes odoo/enterprise#70524

Related: odoo/odoo#181111
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>