[FIX] whatsapp: allow public to send themselves messages
Currently it's not possible for a public user to send themselves a message
over whatsapp when the message contains information linked to a record

This is because most records do not allow public users from reading them.

As whatsapp can only used a select list of "safe" fields we can render the fields
as SU when using the cron.

task-4199766

Part-of: odoo/enterprise#70608
Related: odoo/odoo#195686
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>