[FIX] account_reports: duplicate financial reports
Enable the copy of a financial report. Right now, when duplicating a financial
report (or several at the same time), only the name is copied (same name, not
even a "(copy)" after. We want to improve it by adding this "(copy)" in the title,
and also copy the budget lines. This will enable a user to not start from scratch
every time he wants to create a new financial budget.

task-4207202

closes odoo/enterprise#70805

Signed-off-by: John Laterre (jol) <jol@odoo.com>