[FIX] stock_barcode{,_picking_batch}: more precise subline merge conditions
**Current behavior:**
If a picking has two moves with move lines for the same
lot-tracked product, they will be grouped under one parent line
in the barcode app. Then, if the parent line is selected and a
lot for their product is scanned, it will keep adding the
quantity to the first subline even after it's completed.

**Expected behavior:**
Move lines for the same lot shouldn't be grouped under one
parent if they are for different moves.

**Steps to reproduce:**
1. Create a lot-tracked product and some quant for it

2. Create a new delivery picking and add 1 move to it for the
     lot product -> assign the picking

3. Add another move to the picking, again for the lot product

4. Open the picking in barcode, without expanding the sublines,
     scan the lot created with the quant

5. When the parent line is complete, expand it to see that one
     of the sublines has no quantity

6. Furthermore, validating the transfer at this point will
     create a backorder despite it being technically completed

**Cause of the issue:**
The group key for sublines doesn't take into account the line's
move id.

**Fix:**
Add `line.move_id` to the group key for sublines.

Then, ensure when checking whether to merge some new lot scan
qty with an existing line in `_getNewLineDefaultValues()`, take
into account the lot details (not done prior to this commit).
Additionally, check all available barcode lines for a merge-able
line if there is not one currently selected.

Similarly, check that two barcode lines belong to the same
picking when in a batched context before allowing their
quantities to comingle.

opw-4185767

closes odoo/enterprise#70539

Signed-off-by: Steve Van Essche <svs@odoo.com>