[FIX] planning_contract: check resource is set before checking if flexible
Before this commit, the compute of the allocated hours could be raised a
traceback if the shift is an open one because we check if the resource
is flexible before checking if there is a resource set on the shift.

This commit makes sure a resource exists on the shift before checking if
the resource is flexible or not.

task-4206945

closes odoo/enterprise#70758

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>