[FIX] room: fix frontend view's height
Purpose:
-------
Currently, the room frontend view no longer occupies the full vertical
space of the page.

This issue was introduced by [1].

To fix this issue, the wrapper element (`div#wrapwrap`) is removed from
the room frontend view. This simplifies the DOM structure and allows the
main container to take the full available space.

[1]: odoo/odoo#98429

Task-4207291

closes odoo/enterprise#70541

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>