[REF] knowledge: convert the file behavior to an embedded component
The file embedded component was moved to html_editor. This commits extends the
feature to add the Macro capabilities of the file block, allowing to attach it
to a previously visited record, or to send it by email through that previously
visited record.

task-3672747

Part-of: odoo/enterprise#67083
Related: odoo/odoo#173997
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>