[FIX] stock_barcode_picking_batch: unable to open empty batch pickings
Steps to reproduce:
1. Create a new batch picking (in the Inventory app)
2. Set an operation type and save
3. Open the batch picking in the barcode application

Before this commit:
Because the batch picking has an operation type, Odoo will try to load
it from the cache, but since the batch picking is empty, the backend did
not send any `stock.picking.type`.

After this commit:
Load the `stock.picking.type` in the frontend cache.

opw-4175883

closes odoo/enterprise#71259

X-original-commit: a779aa262b09e79a8601de60379dd3e89121267c
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Walravens Mathieu (wama) <wama@odoo.com>