[FIX] project_enterprise: traceback when adding a new favourite view in the control panel
When saving a new favourire view from the control panel a traceback would appear.

It was because the domain was not parsed correctly in search_model.js when calling the
getDomain method from the _getIrFilterDescription method. It was not parsed correctly because
2 parameters given to the method were lost because the getDomain method was overwritten in
HighlightProjectTaskSearchModel.

After passing the parameters to the super call inside the overwritten getDomain method, the
traceback is fixed.

task-4207512

closes odoo/enterprise#70624

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>