[FIX] mrp: fix mo overview operation cost.
This commit modifies the MO cost of operations.

Now, when the production is started and if there is no duration expected
on the operation, the mo cost of the operation will be the theorical cost
of the real duration (i.e. it will be the result of machine + labour cost for
that duration).

When the production is done, the same behaviour will be applied, but without the
labour cost.
The timesheet cost of an operator for the MO cost column is now computed based on the operation's workcenter employee cost.

closes odoo/enterprise#70854

Related: odoo/odoo#181776
Signed-off-by: William Henrotin (whe) <whe@odoo.com>