[FIX] hr_expense_extract: payment method display name
In this pr: https://github.com/odoo/odoo/pull/174537 we have change the display
name of the payment method line by adding the name of the journal. But it has
changed every view where the payment method line are available.
Since we want to have the journal on the groupby of some view, we added a
context key to be added when we want to hide the journal.

no task id

closes odoo/enterprise#70622

Related: odoo/odoo#181340
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>