[FIX] sale_ebay: log successful API response only in `debug` mode
Description:
Currently the ebay module logs as `INFO` the whole content of the JSON response
coming from the Ebay API. This can bloat excessively the logs of databases,
forcing customers to quickly exceed their disk quotas.
We can just log this type of information as a DEBUG log.

Reference:
opw-4192547

closes odoo/enterprise#70650

X-original-commit: 5f814b33917bf9e85489b7a8fd909a3adcc23b79
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>