[IMP] account_reports: better quarter strings on report
Before this commit quarter strings on reports were not always explicit.
If the fiscal year was not aligned to the calendar year, people would get confused by thinking the quarter was aligned on the fiscal year

Now to avoid that we show the start and the end of the quarter.

task-4200096

closes odoo/enterprise#70647

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>