[FIX] documents{_hr}: avoid giving employee access to employee document
How to reproduce:
- Install documents_hr
- Activate "Human resources" Centralize your employees' documents
- Open "Abigail Peterson" employee form
- Click on the "Documents" stat button
- Upload a document
- Click on the "Share" button

"Abigail Peterson" is set as viewer of the document but mustn't.

Second case:
- Open "Abigail Peterson" employee form, open chatter
- upload attachment

A document is automatically created.
"Abigail Peterson" is set as viewer of that document but mustn't.

To solve the problem, we change the method _get_document_access_ids where we
were giving access to the employee for those cases.

Task-4212817

closes odoo/enterprise#70717

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>