[FIX] web_studio: kanban editor: can toggle bold on field
Before this commit, toggling the bold attribute on a field in a
(new API) kanban view didn't work. This was because the feature
wasn't fully implemented (in web) for new API kanban. Indeed, the
class "o_text_bold" was correctly set, but the associated css
rule was scoped for legacy kanban records.

Instead of fixing it in web, we rather want to promote the use of
bootstrap utility class. So instead of setting the `bold`
attribute, we toggle the `fw-bold` classname.

Task~4199777

closes odoo/enterprise#70669

Related: odoo/odoo#181420
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>