[FIX] l10n_mx_edi: Show reset2draft option for cancelled global invoices
Before this fix, after cancelling a CFDI within a global invoice, the
"reset to draft" option was not visible, preventing the user from
modifying the invoice.

Steps to reproduce:
1. Generate 2 invoices
2. Use the option to generate a global invoice (for any period)
3. Go to any of the individual invoices and cancel the CFDI
4. After the cancellation, the option to reset the invoice to draft is
   not available.

With this fix, the "reset to draft" option will be available after
cancelling a CFDI in a global invoice, allowing users to reset the
invoice for further modifications.

Extra: Was considered in the global invoice test to ensure that the
field to allow reset to draft is True.

closes odoo/enterprise#70697

Signed-off-by: Laurent Smet (las) <las@odoo.com>