[FIX] mrp_workorder: don't show outdated info in shop floor quality view
To reproduce:
- Create MO for 3x Stool
- Open Shop Floor and open the register production step
- Click '+' to generate a serial number
- Continue processing steps until the first register component step

Current behavior:
The step shows quantity to consume 3.
If you close the step and open the step again manually, the quantity to
consume shows 1.

Expected behavior:
The step immediately shows quantity to consume 1 as the qty_producing is
set to 1 after the previous generate serial step.

Part-of: odoo/enterprise#70704
Signed-off-by: William Henrotin (whe) <whe@odoo.com>