[FIX] l10n_fr_reports: show EC Sales List for France
Currently the EC Sales List menuitem is not visible when only installing
`l10n_fr_reports` (and no other localization activating the item).

To reproduce it you need to install `l10n_fr_reports` on a database
where the item is not already visible. (It does not work on the runbot.)

closes odoo/enterprise#71021

X-original-commit: 82d1a385c1bdda9b969963a7b4da3b064ed467e5
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>