[FIX] documents_spreadsheet: Fix document multipage computation
The multipage computation of xslx files was a bit overkill as it would
convert the xlsx file to an o-spreadsheet json friendly form where it
could simply explore the zip files. Furthermore, that convertion would
create attachments for each media file present in the xlsx file.
This is bad in 2 ways:
- the created attachment would never be used again
- we were creating attachment while calling a computed field (read/write
  mix is bad)

Task-4213183

closes odoo/enterprise#75332

X-original-commit: 7ba68ca1a64d10edce0d109e35e7b62da49193ad
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>