[FIX] sale_purchase_stock_inter_company_rules: always allow sync
Previously, it was considered that a company needed to allow the
generation of purchase orders for inter-company transactions in order to
allow the sync between deliveries made to it and its own receptions.

This implied that a company not allowing automated purchases to be made
on itself couldn't sync its receipts with a sale order generated another
company following a purchase to it, which is quite confusing user-wise.

To avoid that, we allow the sync regardless of the selected options for
automated Purchase / Sale Order generation.

Task-4207009

closes odoo/enterprise#70740

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>