[FIX] l10n_de_reports: handle included taxes in datev
Steps to reproduce:
- Install Accounting and l10n_de_reports
- Switch to a German company
- Create a 19% included tax
- Create an invoice with the following line:
  * Product: [any]
  * Quantity: 1.00
  * Price: 19.95
  * Taxes: [the 19% included tax]
- Confirm the invoice
- Go to "Accounting / Reporting / Audit Reports / General Ledger"
- Download "DATEV DATA (ZIP)" and check "EXTF_accounting_entries.csv"

Issue:
The line for the created invoice has an amount of 19.94 instead of
19.95.

Cause:
As the tax is included, the balance on the line is computed and rounded.
When the report is generated, "compute_all" is executed with the rounded
balance to retrieve the total price with the taxes.
Because the balance used in "compute_all" has been rounded and therefore
is not exactly the same than the original computed value, there can be a
0.01 difference.

Solution:
Use directly "price_total" field when available.

opw-4191876

closes odoo/enterprise#71654

X-original-commit: 91ff0f8433871ffbffbbabd598c336d54fc5c4e6
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>