[IMP] web_mobile: lower time_between_reads_in_ms
Lowers the `time_between_reads_in_ms` session setting to 1000 to 100 ms.
The idea is to send more often scanned data.

For now, those data are used only in `stock_barcode` and another batch
is done before processing the data, but it is more suitable to receive
them quicker to update reads count more accurately.

Part-of: odoo/enterprise#70529
Signed-off-by: Steve Van Essche <svs@odoo.com>