[FIX] account_consolidation: disable audit if no journal
Steps to reproduce:
1. Go to Consolidation
2. Create a new consolidation period
    - Add only Sub-Consolidation Periods, no Company Periods
    - You can skip this step on runbot and open the "Top Consolidation"
      period directly
3. Open the Consolidation Balance with the smart button
4. Try to click on an auditable cell
5. You will get an error

The problem arises because the action_open_audit function need a
journal ID. If no journal is defined, the method returns None.
Next, the doAction function in the JavaScript audit method will return
an error.

To fix the issue, I propose ensuring that a journal is defined, and if so,
enabling the audit.

opw-4207691

closes odoo/enterprise#70845

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>