[FIX] documents_spreadsheet: apply revisions in public view
Steps to reproduce:

- (as an internal user)
- create a new empty spreadsheet
- do a couple of changes
- Share the spreadsheet Anyone with the link: Viewer
- open the link in an new incognito tab
- (everything fine at this point)
- now, leave the spreadsheet in the backend
- reload the public spreadsheet page
=> everything disapeared

When the internal user leaves the spreadsheet, a new snapshot is created
and pending revisions are archived.

When the public user reloads, the data comes from the field
`spreadsheet_data` which is the initial data (not the latest snapshot)
and doesn't get any revisions (they are archived)

I also added a test which gets pending revisions.

closes odoo/enterprise#70856

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>