[FIX] helpdesk_sale: fix portal access rights when grouping/searching
Prior to the fix, portal users would get an access rights error whenever
they would group or search sale order.

That was because the ``sale_order_id`` field was only given access to
sale and accounting users. That was the case since the helpdesk app was
first created. All users, including portal, have access to the ``sale_order_line``
field, thus the restriction on the ``sale_order_id`` field are not anymore
necessary.

We "fix" the access right error by basically removing the group restriction
on the ``sale_order_id`` field in the helpdesk.ticket model.

task-4207628

closes odoo/enterprise#70889

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>