[FIX] helpdesk_sale_timesheet: fix invisible group headers when grouping by SO in portal
When grouping by SO in the helpdesk ticket portal list view, the group headers were
invisible.

That was cause by changing the groupby options from ``sale_line_id`` to ``sale_order_id``
without changing the condition in the template that would draw the group headers.

In this commit we change the condition to ``groupby == 'sale_order_id``.

task-4213896

closes odoo/enterprise#70890

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>