[FIX] documents_project: actually prevent deletion of Projects folder
Steps to reproduce
------------------

1. Change the parent of the "Projects" folder to put it in another folder.
2. Delete that folder.

The "Projects" folder will be deleted, despite the constraint preventing it from being deleted.
This is because when a relational field is set as `ondelete=cascade`, we don't call the `unlink`
on the cascaded records and let the database handle it.
Therefore, the constraint is not triggered.

With this commit, we check not only the folder being deleted, but also its descendants.

Task-4241631

closes odoo/enterprise#74263

X-original-commit: 111f2eb5b6cb5b8d7eb0923757e9f88b09133fea
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>