[IMP] mail: restrict template rendering by default
https://github.com/odoo/odoo/pull/164671
offers a way for non-template editors to use simple expressions.

Thanks to this, most of the expressions used in email templates,
such as the partner name in an opportunity email template,
can be used by regular users without the need to be an email
template editor.

Thanks to this, the need to set, by default,
internal users as template editors is no longer a strong necessity.

This revision aims to set the "Restrict Template Rendering" setting
by default.
Companies needing to set everyone as an email template editor
can enable this back by un-checking its related box in the general
settings.

A constraint is added to the appraisal mail wizard to prevent
the rendering of arbitrary templates. UX-wise it changes nothing
as the wizard already does not allow picking the template used.

Refactoring of this wizard should be considered to remove the need
for unrestricted rendering.

task-4246174

closes odoo/enterprise#71619

Related: odoo/odoo#183062
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>