[FIX] web_mobile: avoid using native method if not needed
The `mobile.methods.enableReader()` method requires the Android
Bluetooth permission and so in the current version of Odoo it's asked at
the start of the `WebClient`.
This commit adds a system to enable the loop communication (and so the
`enableReader` method) with the native Odoo Mobile App only when the bus
is actually used.

Note: one test is marked as `skip`, waiting for HOOT's PR that fixes
`expect.step` in `removeEventListener`.

Part-of: odoo/enterprise#70907
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>