[FIX] pos_iot_six: send request with JS
The wizard to configure a Six payment terminal
currently sends a request to the IoT Box via
the server. However, often the DB is not on
the same network as the server so this request
fails.

This PR moves the request logic into a JS widget,
so as long as the client can reach their IoT Box
the Terminal configuration should succeed.

task-4214070

closes odoo/enterprise#71060

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>