[FIX] pos_{restaurant}_preparation_display: retain customer count information
This is the second part of the pr renaming `customerCount` to `customer_count`.

Why the fix:
------------
After the refactoring https://github.com/odoo/odoo/commit/2a5f1abf2e98ee09fa7a912b87d71879b5ff260b,
we use the same field structure as in python. Therefore, when we load/create
a pos order we use `this.customerCount` is undefined since in python the field
is called `customer_count`.
https://github.com/odoo/odoo/blob/ae0b903b753decab1592df09a1e3cf1baa55f566/addons/pos_restaurant/static/src/overrides/models/pos_order.js#L8

By changing `this.customerCount` for `this.customer_count` we not only retreive
the correct field when loading an order in the session but we also send the
correct information to the backend.

opw-4182891

closes odoo/enterprise#71362

X-original-commit: 42da616716851d32d572091964c22e0001c63515
Related: odoo/odoo#182590
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>