[FIX] account_avatax: fix "Tax names must be unique!" errors
We recently reworked the way taxes are named [1]. It considered the
description part of the "key" that uniquely identifies a tax. The
description contains the jurisCode field returned by Avatax [2]. This
key is used to find an already existing tax, and creates it if it
doesn't yet exist.

_constrains_name() of account.tax however, doesn't consider the
description field. This means that two taxes with the same
name (e.g. CA STATE 6%) and different jurisdiction codes (e.g. 001 and
013) cause issues. The first one will be created as normal. When the
second one is processed it will fail to find a match because the
description is different, so Odoo will create a second one resulting
in a constraint error.

To resolve this, stop adding the jurisdiction code in Odoo. This makes
the description identical to the name, so we remove it all
together. In the end it's not critical information that needs to be in
Odoo. For tax details the user already relies on the Avalara portal
anyway.

[1] https://github.com/odoo/enterprise/pull/68549
[2] State assigned code identifying the jurisdiction. Note that this
is not necessarily a unique identifier of the jurisdiction.

closes odoo/enterprise#70980

Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>