[FIX] l10n_eu_oss_reports: branches xml report export
Description of the issue this commit addresses:

Branches companies are not allowed to export to XML the OSS reports while
this is a use case that should be allowed for specific scenarios.

Desired behavior after this commit is merged:

The XML export button is usable by any branch company.

closes odoo/enterprise#71320

X-original-commit: 67dec08b129f894fb18b84a06f2c430ddec260d8
Related: odoo/odoo#182508
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>