[FIX] l10n_ch_hr_payroll_elm: Make IJM and LAAC Position Agnostic
Currently, order of Many2many insurance_line_ids matters for IJM and LAAC, which can be error prone in the case
where you go from example from 2 solutions -> 11,12 to one solution -> 12, the cumulated salaries
will not be computed properly in this case. We make the rules Position aware in this PR, meaning that
order or quantity of solutions in the contract does not matter anymore.

closes odoo/enterprise#71012

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>