[FIX] website_knowledge: fix login redirection url for view links
When a visitor clicks on a view link, they are redirected to the login
page to authenticate and gain access to the view. This is necessary to
ensure the user has the appropriate permissions to open the view and
access the records.

Current problem:

When clicking on a view link, the users are redirected to an error page
after successfully logging in. This happens because the system generates
a faulty redirection URL (`/knowledge/article/undefined`) for the login
page. This occurs because the `KnowledgeWidget` widget generates the URL
using `this.resId` (`/knowledge/article/{this.resId}`) and the widget
does not have any `resId` property set.

Steps to reproduce:
1. Open any list or kanban view
2. In the view's cog menu: "Knowledge" > "Insert link in article"
3. Click the "New" button to create a new article
4. After being redirected to the article, click on the "Share" button
5. Publish the article
6. Copy the generated article URL
7. Open a private window and paste the URL
8. Click the view link
9. Enter valid credentials
=> After logging in, the user is redirected to an error page.

TO BE: After logging in, the user should be redirected to the backend view
of the article. The user should then be able to open the view by clicking
on the view link.

Solution:

To solve this issue, we will use `this.$id` instead of `this.resId` to
get the current article id.

Reference: https://github.com/odoo/enterprise/pull/57346

task-4210743

closes odoo/enterprise#70651

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>