[FIX] documents: fix breadcrumbs for non-pinned Company roots
To reproduce:
Open a non-pinned folder in "Company", the breadcrumb will not be correctly
shown and actions are not available.

When simply removed from the returned records, no item is added in the
breadcrumbs on opening a non-pinned company root folder.

Tests are disabled until we fix them with the others.

Task-4216195

closes odoo/enterprise#71042

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Co-authored-by: Lopes Marc (loma) <loma@odoo.com>