[FIX] sign: display saved message only if active template
Change
-----
UX improvement: a "Saved" message is displayed when leaving the template
only if it is active, as it can be confusing for the user to not be able
to find it (archived records are not visible by default).
This happens when uploading a pdf as a document to sign.

opw-4207093

closes odoo/enterprise#71451

X-original-commit: dd5ff516ae3be271f8213cf998c2cf36f03e5856
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>