[FIX] appointment_account_payment: blacklist product service tracking
In the Appointment app, when trying to select an "Up-front payment product," products configured to create a task in a project are not available in the dropdown list.

Cause of the issue:
In the `appointment_type` model definition within the `appointment_account_payment` module, a domain was applied on the `product_id` field, restricting the selectable products. The domain filtered out products with the `service_tracking` field set to values other than `"no"`. As a result, products that create tasks in projects (which have `service_tracking` set to `"task_global_project"`) were excluded. The commit https://github.com/odoo/enterprise/commit/ff54ff6 aim was to prevent `event`, `event_booth` and `course` product to be selectable. This commit allows other product whose `service_tracking != 'no'` to be selectable while not allowing `event`, `event_booth` and `course`.

opw-4146731

closes odoo/enterprise#73092

Related: odoo/odoo#185903
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>