[FIX] knowledge: fix attachment display in comments
This commit aims at fixing an issue with message containing attachments. Those
attachments can be displayed too small resulting in scrollbars and weird
displays.

In e8df9ed4a92e we tried to fix this by using the environment variable
inChatWindow, but this changed the styling of comments drastically. So in this
commit we remove this option and directly apply the correct style to the
attachment cards.

task-4221276

closes odoo/enterprise#71100

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>