[FIX] account_batch_payment: sign of batch payments
The sign of the batch payment should depend on the direction of the
payment when there are no journal entries, just like when there are
journal entries.
In the case there is one, we were taking the `amount_currency` (which is
direction dependent), but not in the other case.

Part-of: odoo/enterprise#70778
Related: odoo/odoo#181733
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>