[FIX] appointment: fix unnecessary action reloading
Steps to reproduce
==================
- Got to appointment type;
- Apply some filter.
- Now create a share link.
- Filter is lost.

Technical
===========
https://github.com/odoo/enterprise/pull/63046/commits/492296e91e9c0b1106a7bc6852010be426d2abe8
Here we are loading the action while creating or updating the share link
to reflect the changes without refreshing.
It is not required when we create a share link from other places instead of a
Share link menu.

After this commit
==================
This commit addresses the issue and now the filter will not lost when we create
a share link from appointment_type.

Task-4138356

closes odoo/enterprise#71165

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>