[FIX] account_reports: fix residual computation in bank reco. report
There are currently these issues in the bank reconcilation report
1. There may be a traceback due to a division by 0 error
2. The residual amounts are not correctly computed in the report.
   It always displays the full amount; even if the move is partly
   "reconciled" already.
3. Duplicate entries can appear.

All issues are solved by rewriting the generation logic a bit.

Issue (1) i.e. happens due to move lines with a 0 balance in the move
associated with the bank statement (line).

All 3 issues can happen due to the way the entries in the report are
generated:
Consider a single bank statement (line) / its associated "bank move".
To generate the line the report we join the bank statement line and
bank move to each line of the bank move (except the liquidity line).

The report groups by move line id. This generates 1 row in the
report per move line (except the liquidity line) of the bank move.
This directly causes issue (3).
It also causes issue (2) due to the following:
The reason we join all the move lines in the first place is to
compute the unreconciled amount of the bank move. (By summing amounts
on the move lines). But when generating a single row of the report
we only have the information about 1 line (due to the grouping by id)
and not the whole sum. Thus we do not have the full residual information.

opw-4178134 (Issue 1)

closes odoo/enterprise#71243

Related: odoo/odoo#182589
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>