[IMP] account_base_import,l10n_se_sie_import: prepare for SIE 4
This commit cleans the code related to some of the implemented SIE
import in preparation for adding the new SIE 4 module.

From now on, "SIE import" will be called "SIE 5 import" to differentiate
between version 4 and version 5. In master, the module will be renamed
to l10n_se_sie5_import to further make it more clear.

In addition to the text changes, the placement of the button(s) for SIE
import is also refactored to make it cleaner to extend by the SIE 4
module.

Task-id: 3735808
Part-of: odoo/enterprise#66263
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>