[ADD] l10n_se_sie4_export: Export data to SIE 4 files
This commit adds a new module for exporting accounting data in General
Ledger to a SIE 4 file.

To be able to export, the selected date periods must be of type fiscal
year, and exporting multiple companies data at once is forbidden.

This first implementation does not at all use the functionality of the
"#OBJEKT" and "#DIM" items.

SIE 4 documentation:
https://sie.se/wp-content/uploads/2020/05/SIE_filformat_ver_4B_ENGLISH.pdf

closes odoo/enterprise#80403

Task-id: 3735808
X-original-commit: b8ba408774532ffd1ff00a0fcd78f8847ad008c7
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>