[FIX] account_accountant: always display other lines in bank reco.
Consider an unreconciled bank statement (line) in the bank
reconciliation widget.
Currently only the liquidity and suspense/transfer line of the
associated move are shown.
This means that some lines of the move are not shown.

After this commit we show all the line (so i.e. the other lines).

Since the diff is not the nicest:
The commit basically just removes the outer `if is_reconciled`.

closes odoo/enterprise#71504

X-original-commit: 405433f9e0598a8ba660c418592e1344edce5bc3
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>