[FIX] industry_fsm_sale: catalog view used in unexpected places
For all 17.0 databases of a specific partner (strangely it couldn't be reproduced on runbot),
the fsm catalog view was used to select product on sale order lines in mobile view.

This commit ensures that this custom view will only be used for the specific action using it
and won't appear in other places.

opw-4196351

closes odoo/enterprise#71300

X-original-commit: 647806cb89f219f5c48a20686e6f52bf1dae0af8
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>