[FIX] industry_fsm_sale: catalog view used in unexpected places
For all 17.0 databases of a specific partner (strangely it couldn't be reproduced on runbot),
the fsm catalog view was used to select product on sale order lines in mobile view.

This commit ensures that this custom view will only be used for the specific action using it
and won't appear in other places.

opw-4196351

closes odoo/enterprise#71297

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>