[FIX] delivery_sendcloud: fixed price datatype when None
Before this commit:

In a recent commit to allow sendcloud to generate labels without having to fetch rates.
The returned value for field `price` is `null` or of datatype `str` when it should be `float` for correct sorting.

After this commit:

Added a check when setting price to make sure correct datatype is set.

opw-4222636

ID of the commit mentioned above: `1f4d6ef5a841fda7c98caa1e74681800898cb30a`

closes odoo/enterprise#71367

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>