[FIX] planning, sale_planning: added recurrence deletion in gantt popover delete
As of now the gantt popover delete button is added through gantt-popover
template which is making it harder to extend its functionality.

In this PR we move the button to the ganttPopoverProps. We can pass the button
through JS side, thus being to extend it's functionality of reccurence deletion.

The unschedule gantt popover button is migrated to JS side to preserve the order
of buttons in gantt popover.

task-4224792

closes odoo/enterprise#71759

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>