[IMP] spreadsheet_dashboard_sale_subscription: fix error while selecting users
Currently, an error was generated when the user tries to filter
`Customer/Saleperson` in the subscription dashboard.

error: `Invalid field sale.subscription.report.message_partner_ids in
leaf ('message_partner_ids', 'in', [3])`

This is because we have used fields 'message_partner_ids' and 'activity_user_id'
to filter out records, but this field is not available in the model.

This commit will fix the above issue by using fielels 'partner_id' and
'user_id' to fielter records.

sentry-5657224204

closes odoo/enterprise#71360

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>