[FIX] helpdesk_repair: don’t link repair moves to the return
Steps to reproduce:
- Create a Helpdesk team with the options "Return" and "Repairs"
- Create a product, sell 1 unit and deliver it.
- Create a ticket for the Helpdesk Team, select the customer who bought
  the product > create and validate a return of the product to your
  warehouse through the ticket.
- Create an associated repair order and add a line for an other product
  in the part tabs
- Go back to the return picking

> the product added to the repair was also added to the return picking

Cause of the issue:

A `default_picking_id` is added to the context of the form view of the
repair order by the `_prepare_repairs_default_value` of the
`action_repair_order_form` so that it prefils the form of the repair:
https://github.com/odoo/enterprise/blob/e74b05fa7a777053fd0e2cd5a265de365e1e434a/helpdesk_repair/models/helpdesk_ticket.py#L36-L40
https://github.com/odoo/enterprise/blob/e74b05fa7a777053fd0e2cd5a265de365e1e434a/helpdesk_repair/models/helpdesk_ticket.py#L55-L56
However, this values stays in the context and will be added as a default
value for each of the move that will be created from this view.

Fix:

Thanks to commit odoo/odoo@3898d61046cc0338082fdcbe0df282241372361a, we
now havea default getter that can be used to associate a default picking
id to the repair order when necessary without the need to put a
`default_picking_id` in the context.

opw-4159779

closes odoo/enterprise#71337

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>