[FIX] l10n_au_aba: handle special characters in ABA file
Steps to reproduce:
- Install l10n_au_aba
- Switch to an Australian company (e.g. AU Company)
- Configure batch payments according to fiscal localization docs in Odoo
- Create a vendor with a special character in the name (e.g. a curly apostrophe "’")
- Configure a bank account for the vendor
- Create a bill for that vendor
- Confirm the bill
- Register a payment with "ABA Credit Transfer" as payment method
- Go to the payments list
- Select the created payment and create a batch payment
- Validate it
- Download the generated ABA file
- Try to validate the file on an online ABA validator
(e.g. on the BCU Australian website)

Issue:
The online ABA file validator finds an error because the accepted length
for a row is not respected.

Cause:
Each row information is encoded with a specific number of characters
where blank characters are added to reach that length.
In this case, the name of the account holder is containing a special
character that corresponds to 3 characters when converted to utf8 and
too many blank characters are added.

Solution:
Compute the length in utf8 before adding the blank characters.

opw-4059837

closes odoo/enterprise#71580

X-original-commit: 99331ab958fc2fc7f6d506448c8f8a2f6a1de84a
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>