[FIX] hr_payroll: no traceback adding line to draft payslip wizard
Steps
---
* Create a payslip for an Employee with a contract (e.g. Anita Oliver)
  (*Payroll > Payslips > All Payslips > New*)
* From the *Edit Payslip Lines* action, try to add a new worked day line
  (first list) with a non 0 amount
* *Validate Edition* => Traceback

Cause
---
When saving new worked days lines, we try to recompute the amount of the
*Basic Salary* payslip line, but if the payslip is still a **draft**
it was never computed, and so the action we are calling does not exist.

Furthermore, when updating, we call the orm to recompute payslip lines,
but the framework tried to prevent us because the field's component was
destroyed/reset.

task-4191349

closes odoo/enterprise#72031

X-original-commit: 19d51e58a9df05e25525b6cd638bf20b67becce7
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>