[FIX] hr_payroll: no traceback adding line to draft payslip wizard
Steps
---
* Create a payslip for an Employee with a contract (e.g. Anita Oliver)
  (*Payroll > Payslips > All Payslips > New*)
* From the *Edit Payslip Lines* action, try to add a new worked day line
  (first list) with a non 0 amount
* *Validate Edition* => Traceback

Cause
---
When saving new worked days lines, we try to recompute the amount of the
*Basic Salary* payslip line, but if the payslip is still a **draft**
it was never computed, and so the action we are calling does not exist.

Furthermore, when updating, we call the orm to recompute payslip lines,
but the framework tried to prevent us because the field's component was
destroyed/reset.

task-4191349

closes odoo/enterprise#73326

X-original-commit: 06c647091aa3fb8d0952caa3e4ad6089b7e413a4
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>