[FIX] l10n_es_reports: change Q4 fields to 0 instead of empty
According to the documentation of the new Modelo 303 BOE export, the new
fields that will be used for the declarations starting in 10/2024 or
Q4/2024 "can only be completed from periods 10 and 4Q of 2024 and
subsequent years". As such, we left the fields "empty" using the space
character (like in other places).

However, now that the new AEAT submission page is ready, it seems they
expect us to fill the new fields with zeroes instead of leaving them
empty. This commit fixes that.

opw-4222842

closes odoo/enterprise#71407

X-original-commit: 0838ddab96228c9e3a88152a9d0832e384244741
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>