[FIX] account_disallowed_expenses: fix audit of categories with vehicle split
Currently, when auditing a DNA category with the vehicle split option leads to
incorrect values. All the lines from the account being audited are displayed,
including those with a 'vehicle_id'.

That is wrong, since these lines will be grouped with the corresponding vehicle
inside said category. They will appear both when auditing the category as well as
when auditing each vehicle from  this category.

The solution is simply to exclude lines related to a vehicle when auditing a
category while the vehicle split is active.

task-4023543

closes odoo/enterprise#71758

X-original-commit: 0bc2a06edbe6eb57b794eb7db5a4b8019a48132c
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>