[FIX] project_timesheet_forecast_sale: convert project report to product uom
Steps to reproduce:
- Create a product with the following settings:
-- Product Type: Service; Invoicing Policy: Based on Timesheets
-- Create on Order: Project & Task; Unit of Measure: Days
-- Sales Price: 120 (Use != 100 when using Mitchel Admin)
- Create a quotation for this product and confirm it
- Created Task > Timesheets tab > Record 4 hours
- Go to Planning > New > Create a 1 day shift for Mitchel Admin
- Link that shift to the project created by your SO
- Enable 'Project Planning' in Settings > Planning
- Go to Project > : Menu > 'Timesheet and planning analysis'
- Under the Measures dropdown, add all columns

You will notice that the measures for revenue and margin are computed as
though the unit price corresponds to 1h even though our product uom is
Days. This makes the computations for revenues and margins completely
wrong. This happens because the module converts all values into hours
without accounting for the fact that the SOL unit price is not
necessarily by the hour. This happens regardless of the 'Encoding
Method' setting of Timesheet.

opw-3918082

closes odoo/enterprise#71425

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>