[FIX] documents: fix multicompany user permission for admins
Reproduce:
1/ Create documents and assign a company an admin (`group_erp_manager`)
   doesn't have access to
2/ As that admin, open the Documents app
3/ Go to "Home" and expect an Access Error

There was a mismatch between `user_permission`'s search and compute
methods for admins.

We also take the opportunity to fix the SHARED (and clean COMPANY)
search panel `CategoryDomain`s as odoobot's ownership wasn't used
to make documents visible in COMPANY (therefore not in SHARED) on its
own.

A test is added.

Task-4241620

closes odoo/enterprise#71462

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>