[FIX] appointment: only show previous date slots in month
TO REPRODUCE
===========
1. Create an appointment with users A and B
2. ...using 'no picture' and 'user then time'
3. go to front end
4. pick a date D both users have slots in
5. go to next month
6. change user in the dropdown (or change TZ)
7. slots appear for day D in previous month

ISSUE
=====

When selecting an other resource or user in the dropdown in
resource_time mode, even if the selected one has no available
slot for the current month, slots will appear in the slot list,
corresponding to the previous date, if the new user has slots
there. Even if that date is in another month! The same happens
when changing tz in the dropdown.

This leads to a very strange mismatch between the calendar
dates and the slots. The user may select a slot and book for
a date in a month they do not currently see on the calendar.

SOLUTION
========

Now, we will select the previously selected date if it is in the
current month and has slots. Otherwise, we let the user pick their
date manually.

Task-4169513

closes odoo/enterprise#73673

X-original-commit: 520f78ade72c210ec4d934d4eddbfe963ddba522
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>