[FIX] helpdesk: Correctly compute 7-day SLA success rate
The SLA success rate was previously being computed
using the entire history of closed tickets as the
domain. This is contrary to the expected behavior,
given the helper text on Odoo and the list view
of tickets that opens up if you click the
percentage.

This commit changes the domain to only consider
tickets that were closed in the last 7 days.

To reproduce on Runbot, follow the steps laid out
here, as the process involves forcing tickets to
fail and be older than 7 days closed:
https://www.odoo.com/odoo/49/tasks/4206255

opw-4206255

closes odoo/enterprise#71600

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>