[FIX] mrp_workorder: update MO count and stop timer
There's no refresh when closing a MO on the 'All MO' tab of the shopfloor.
Because of that, the MO fades away but the total count is still the same.
This PR forces a refresh to make sure that the count is corresponding to
the number of MOs shown.

When marking a workorder as done with the timer running, the timer
doesn't stop and the value registered contains the time it takes to
fade away.
E.g. if you mark a WO as done at 1 min 33 sec, the real duration
registered will be around 1 min 36-37 sec.
This fix ensures the timer stop effectively and visually when closing
a WO.

closes odoo/enterprise#71695

X-original-commit: 088d754b3f538eef3257479a4e053f3af965f12f
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>