[FIX] hr_payroll_expense: raise error on missing expense salary rule
Steps to reproduce:
- Expenses > New > Mitchell Admin; Paid by employee (to reimburse)
- Create report > Submit to Manager > Approve > Report in Next Payslip
- Payroll app > Payslips > To Pay > New > Mitchell Admin
- Set structure to 'Worker Pay' (Or any other with no expense rule)
- Compute sheet > Create draft entry > Journal entry (Draft) button
- Post > Traceback

The traceback is caused by 'KeyError expense_report_name', this happens
because e65511ac69131b048da5d11afd00f3cfe91db474 changed the move
creation process to ease reconciliation of expense compensated on the
employee's payslip. We would have previously had access to
expense_report_name=move.expense_sheet_id.name, which we do not in
saas-17.4 and onwards. This causes the traceback when trying to display
the error message which still depends on this inaccessible variable.

We instead expect expenses not to be linked when creating payslips with
a structure containing no expense rule, and a proper error if the user
attempts to add an expense by hand on such a payslip.

opw-4222515

closes odoo/enterprise#72514

X-original-commit: 6e4a94f530643a9ad7b461d49e149a339bc6a341
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>