[FIX] l10n_mx_edi: calculate equivalenciadr with currency rate
Steps to reproduce the issue:

1. Activate Mexican localization
2. Create multiple Invoices in USD with different amounts (add the Invoice Date)
3. Send and Print them for the CFDI
4. Register a grouped Payment in MXN for those invoices
5. Go to one of the Invoices and click on Update Payments
6. Go to the CFDI tab and, on the Payment line, click on Force CFDI
7. On the Payment line, you should now be able to Download an XML document
8. In the document, the EquivalenciaDR attributes should have the same values, which is not the case

Explanation:

When calculating the rate (which then is used in EquivalenciaDR) in `_l10n_mx_edi_add_payment_cfdi_values`, we are using amounts that have been rounded out, leading to rates that differ when aiming for certain decimal precisions.

Fix reasoning:

While `_get_conversion_rate` can be used to avoid the inconsistencies of the rounded amounts, it should not be used if the values do not match once they are converted, since the rate would then be wrong.

opw-4090269

closes odoo/enterprise#70083

Signed-off-by: Laurent Smet (las) <las@odoo.com>