[FIX] account_online_sync: remove readonly for auto_sync
A previous commit (https://github.com/odoo/enterprise/pull/63926/commits/c749e1358dbd0175d0de5f32eab2c4dd98053905)
added a readonly condition on the field auto_sync of the account.online.link
view. This was made to prevent the automatic fetching of transaction from
interactive providers when the connexion had expired.

The issue lies in that condition being based on the field it applies to. This
does not work well with manual editing as the readonly condition would update
before  hitting 'save', preventing to save the changes when unticking the
'auto_sync' checkbox.

After discussion with FLG, we decided to keep things simple and remove that
condition altogether. Else, one would have had to create a new field and import
new data from the instituion (is_interactive).

No opw but the issue was raised in a odoofin support discord thread.

closes odoo/enterprise#71643

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>