[FIX] mrp_workorder: display additional component on shopfloor
Steps to reproduce:
- Create a product with a bom and an operation op1
- Create and confirm an MO for 1 unit of that product
- Go to the shopfloor > on the operation > 3 dots > Add component
- Add a component to be consumed in that operation

> While the componenet was correctly added to the MO and linked to
> the WO it is not visible on hte shopfloor.

Cause of the issue:

The only moves that are visible on "MrpDisplayRecord"'s are the move
that are in `manual_consumption = True` see:
https://github.com/odoo/enterprise/blob/88a092acd261db8f3a8ee1f0cf8174c3d7089472/mrp_workorder/static/src/mrp_display/mrp_display_record.js#L188-L191
However, when a product is added via the `action_add_component`
> `add_product`, the `manual_consumption` of the move is computed as if
it was added to the MO without being linked to a `bom_line_id`:
https://github.com/odoo/enterprise/blob/88a092acd261db8f3a8ee1f0cf8174c3d7089472/mrp_workorder/wizard/additional_product.py#L36-L46
https://github.com/odoo/odoo/blob/2d652f42f63c2501b1a9feb7a4830ea3d7d05f51/addons/mrp/models/mrp_production.py#L1198
As a result the move_raw is added as a manual_consumption = False.

Note:

Thank to Commit eb751b5f861b25205dbb0b0fb90ddcdbbcd10f55 , the issue is
not reproducible in 18.0 anymore. This was soved in the same manner:
https://github.com/odoo/enterprise/blob/2a0ccbe7058f8d4383a63ed1897945bec3a1cec3/mrp_workorder/models/mrp_production.py#L36-L38
https://github.com/odoo/enterprise/blob/2a0ccbe7058f8d4383a63ed1897945bec3a1cec3/mrp_workorder/models/mrp_production.py#L107-L114

opw-4199156

closes odoo/enterprise#71803

Signed-off-by: William Henrotin (whe) <whe@odoo.com>