[FIX] helpdesk_timesheet: avoid merging helpdesk ticket timesheets
Steps to reproduce:
- Helpdesk app > Any team > : Menu > Enable 'Timesheets' setting
- Timesheets app > Start timer
- Set your helpdesk project and any task (Do not add a description)
- Stop timer > Do it again with same the project and a different task
- Stop timer > Time is added to previous timesheet

The expected behavior would be to create a new timesheet since the task
is different. The reason both timesheets are merged is that we check
'task_id' for redundance while helpdesk 'tasks' are instead recorded in
the 'helpdesk_ticket_id'. Since it is forbidden for a timesheet to have
both a 'task_id' and 'helpdesk_ticket_id', it is guaranteed that task_id
will be False for both of our entries, leading to a merge. Similarily we
can use that property since adding helpdesk_ticket_id to the domain
won't change the usual behavior (the field will always be False if the
project is not from Helpdesk) to add an extra condition here. Source: we
have a contraint '_check_no_link_task_and_ticket' which ensures that.

opw-4203678

closes odoo/enterprise#71784

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>