[FIX] spreadsheet_edition: process aknowledge message directly
With this commit, the user will not have to wait for the acknowledge
message coming through the websocket to process it, it will be processed
directly after the message is sent, if the message is accepted server-side.

With this commit, we are now robust to the case the websocket connection
is dropped **and** the user is alone in the spreadsheet.

closes odoo/enterprise#71981

Task: 4243943
X-original-commit: 95b3a9de5d4fcc89289c34d8df8869c7494b82b2
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>