[FIX] spreadsheet_edition: process aknowledge message directly
With this commit, the user will not have to wait for the acknowledge
message coming through the websocket to process it, it will be processed
directly after the message is sent, if the message is accepted server-side.

With this commit, we are now robust to the case the websocket connection
is dropped **and** the user is alone in the spreadsheet.

closes odoo/enterprise#72246

Task: 4243943
X-original-commit: 47fc31b82dbed637d8b9e1237079e86305d5a15b
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>