[FIX] account_reports: partner filter traceback
When the partner filter gets a lot of partners removed one after the
other very fast, sometimes the filter update request fails, and
the `this.data` variable becomes `null` while waiting for data.

When this happens, we just ignore the invalid data to mitigate the problem.
In saas-17.2 this has anyway being amended by introducing a delay between
the partner cancellation and the the filter update. Subsequent
requests just reset the delay, so one only request is done in the end.

opw-4020313

closes odoo/enterprise#71792

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>