[FIX] l10n_{be,hk}_hr_payroll: Set leave type records as noupdate
The definition of these records include the field `requires_allocation`,
so updating them will trigger the constraint for that field.

There's already a fix for this in https://github.com/odoo/odoo/pull/182533 in 16
but it will be interesting to still mark these records as noupdate to be
coherent with all the other `hr.leave.type` records.

To reproduce:
- Install `l10n_hk_hr_payroll` with demo data
- Upgrade the module

closes odoo/enterprise#71948

X-original-commit: a5684bfb67eba4b56b65be580efec892d11ba48b
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Agustín Castro Bugallo (agcb) <agcb@odoo.com>