[FIX] spreadsheet_edition: remove useless data given to spreadsheet
Before this commit, the data given to the spreadsheet init callback in
the case of inserting a pivot view was not used. This could lead to
an error if the data was too big to be stored in the browser's session
storage.

This commit removes the useless data given to the spreadsheet init.

Note that it's not necessary to write a new test for this, as the
existing tests already cover this case.

closes odoo/enterprise#71916

Task: 4255049
X-original-commit: 402228f94323a4cc1b74a5593e21e457d88b51d1
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>