[FIX] account: wrong waiting and late count on dashboard
Description of the issue this commit addresses:

Since recently, the <[X] To Pay> and <[X] Late> buttons don't send the user to
an account.move.line model view anymore but to an account.move model one but
when that change was made, the computation of the number of items on the
dashboard was not changed so it was still counting the amount of account.move.
line there was in the account.move items that were shown in the view resulting
in wrong totals.

Steps to reproduce:

1. Install account
2. Create new Vendor Bill with a split payment term ("30% now, Balance
        60 Days" for example) on today's date for Bill Date.
3. Go to the dashboard, click the <[X] To Pay> button.
4. the amount of moves in the view that is opened with the button is one
        above the value of "X" in the button.

This is due to using a split payment term that creates two installment for a
single vendor bill hence counting one more aml than there are moves.

Desired behavior after this commit is merged:

The right amount of To Pay and Late moves is shown at all times.

Note on the fix:

The fix for this issue is in the community PR of the bundle, the fix modifies
a method which was overriden in enterprise and this PR addresses the override
to match with the new version of the method.

no task

closes odoo/enterprise#70725

Related: odoo/odoo#181535
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>