[FIX] mrp_workorder: don't set qty for component...
...added from catalog in the Shop Floor.

Description
===========

Before this commit, the quantity of a new move raw added through the
catalog in Shop Floor was set directly in
`_get_new_catalog_line_values`.
With this commit, it is not the case anymore, that way, if we should
pick the component (production in 2 steps), the picking will be
correctly created (before this fix, the picking was created only after
an updated by `_get_new_catalog_line_values`.)

How to reproduce
================

- In Inventory settings, enable "Multi-Steps Routes";
- In your warehouse configuration, select the 2 steps options for the
  manufacture;
- Create a product with a BoM with at least a component;
- Create a second product component (but not as a part of the BoM;)
- Add quantity in stock for both components;
- Create and confirm a MO for the product with the BoM;
- Go in the MO's transfer and validate the "Pick Component" picking;
- Go in Shop Floor, and for this MO, add the second component: click on
  the gear icon, "Add Component", search your component and click on it.

If you check your MO in the back end, you'll see no "Pick Component"
transfer was created for the added component. That said, if you add
more quantity for this component (still through the Shop Floor catalog),
you'll see the transfer will be created but 1 qty will be missing.

Explanation
===========

When in 2 steps manufacture, if a component line is added and the demand
is greater than the actual quantity, a Pick Component transfer will be
created for this product.
In the catalog, when using it through the Shop Floor, the raw move is
created with an equal quantity and demand, which means no Pick Component
transfer is needed.
That said, when updating an existing raw move through the Shop Floor
catalog, the demand is updated BEFORE the actual quantity, resulting of
the creating/update of the Pick Component transfer.

To fix this issue, the quantity is not set directly during the move
creation but will be updated just after (see odoo/odoo#183532)

closes odoo/enterprise#71791

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>