[PERF] spreadsheet: avoid useless encoding and decoding
This commit removes unnecessary base64 encoding and decoding operations when
computing `spreadsheet_data` which contains the serialised json data.

Previously, the binary field `spreadsheet_binary_data` was base64-encoded (as is
the case with all binary fields), meaning the file was read, encoded to base64,
and assigned to `spreadsheet_binary_data`. To retrieve the serialized JSON, it
then had to be decoded again.

By avoiding this redundant steps, with a spreadsheet file of 16.7Mb,
we optimize the time spent in `_compute_spreadsheet_data`:

closes odoo/enterprise#71879

Before: ~100ms
After: ~30ms
Task: 4256155
Related: odoo/odoo#183556
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>