[FIX] l10n_mx_edi: Addenda autozone error for invoice with sale order
This commit targets to fix two errors:

1. With an invoice that has a sale order related to it, when using
the Addenda Autozone a traceback appeared.

In attribute t-att-PODATE for the Autozone Addenda we are sending a datetime object
to the strptime function of datetime which in reality it needs to receive a string
object instead. This raises a TypeError

Since the objective is to set a string date, we convert the date_order into a string date
using strftime function from datetime module

- Install 'Sales' application and 'l10n_mx_edi' module
- In a company with mexican localization selected go to Sales > Quotations and create a new Order
- Select a contact with the Addenda Autozone selected
- Confirm it and create the corresponding invoice
- Confirm the invoice and generate the CFDI

2. Traceback with invoice that have multiple sale order linked.

In attribute t-att-PODATE for the Autozone Addenda is not expecting a recordset when initializing the
value of sale_id, this causes a ValueError: expected singleton error when retrieving value of date_order
when we create a invoice for multiple sale orders

Initialize sale_id with the first order retrieved to use it

task-no

closes odoo/enterprise#71990

X-original-commit: bf0792c360272cf8a345ac631e9dc7413b18133a
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>