[FIX] account_reports: only change the tax lock date when posting the domestic closing
Before this commit, posting any tax closing would always set the tax lock date. This was a bit restrictive, and required making lock date exceptions in order to post operations for other non-closed reports. We alleviate that by only setting the tax lock date when the domestic closing is posted ; the idea being it's the main one. So, tax closings posted for other countries for which a foreign VAT fiscal position is defined do not affect the tax lock date anymore.

closes odoo/enterprise#71896

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>