[FIX] {pos_restaurant_}appointment{_google_calendar}: interface issues
1) Action button on Appointment Type kanban view only appear on hover,
it should be always visible.

2) When a user books a meeting, Odoo creates an activity, which is unnecessary
for calendar event linked with appointment type. This commit exclude appointment
model in valid_activity_model_ids using helper function, so we won't get
any activity on appointment type. (See community commit.)

3) The 'Connect Google' button with info and warning it's too vague for users.
Add new warning when there are no creds but Google Calendar is Installed.
Show old warning + info if creds are found but not synced.

4) The translation icon is badly positioned, it should be inside the field.

5) On Appointment Resource Create, Make Timezone field invisible if there
is no resource_id.

6) When creating a booking in POS, the admin is always set as an attendee,
becuase default for `default_partner_ids` was set to current user on create,
which is not correct.

7) Change booked_mail_template_id field's name to 'Booking Email', as it may contain
mail for appointment event which yet to be confirmed.

Task-4243603

closes odoo/enterprise#72274

Related: odoo/odoo#184272
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>