[FIX] account_reports: Fix expand_groupby for non-relational fields
When doing a groupby on a non-relational field (e.g. a char), if the
field contains NULL values in DB, they appear as None values in the
keys of the `group_lines_by_keys` dict.

If the comparison operator is not implemented between the field type
and NoneType, `sorted` will fail with a traceback.

Solution: sort on keys `(k is None, k)`.
if k is None, then `k is None = True`. Because `True > False`, these
values are pushed to the end by the sorting.

task-none

closes odoo/enterprise#72160

X-original-commit: d279038fe0b9a5a0c613ecd2de81f5156b6fe9f3
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>