[FIX] account_reports: show non-account names again on xlsx export
After [this commit], there was an issue for reports where we don't have
accounts as the first column. It would offset the cell too far to the
left, assuming there is an account code column.

This commit fixes that, so reports like the partner ledger are exporting
correctly again showing the partner names.

[this commit]: https://github.com/odoo/enterprise/commit/a3d76fc2b39428b7ea7c41bf98b2ecbe741a67fb

closes odoo/enterprise#72134

X-original-commit: b2efd2e3d844d84a594d39a4eaa38eb39451e492
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>