[FIX] delivery_sendcloud: update parcel shipping name if necessary
Steps to reproduce:

- Create  and configure a Sendcloud shipping method
- Create a Contact Bob
- Create a subconcat of type delivery for Bob without setting any name
> by default its display name will be "Bob, delivery address"
- Create a sale order for Bob
- Add a Sendcloud shipment method to that SO and confirm the SO
- Go on the associated delivery and try to validate it

Expected behavior:

You should raise a user error telling you that you a pickup point was
required for the selected shipping method.

Current behavior:

You raise a traceback prior to that point.

Cause of the issue:

Since you did not set a name on your delivery partner the code will
break as soon as you try to generate the parcels data because "False"
is not iterable, see:
https://github.com/odoo/enterprise/blob/37577535a4482eaa1250b41136b47b3613eb7693/delivery_sendcloud/models/sendcloud_service.py#L576-L577

Fix:

If for some reason the name of the delivery partner has been removed
or was unset (e.g. display name is of the form: Bob, Delivery address),
the name of its parent should be used to generate the parcel data.

Note:

This fix should normally not be that relevant, because, if you were to
select a pickup point, the partner used to generate the parcel would
already be the parent partner and is guarranteed to have a name:
https://github.com/odoo/enterprise/blob/37577535a4482eaa1250b41136b47b3613eb7693/delivery_sendcloud/models/sendcloud_service.py#L556-L561already
However, now, if you proceed with this flow without any pickup point
(which you can only select form the website) you will now raise a
relevant user error:

Invalid Operation:

"A service point is required for the selected shipping method"

opw-4118362

closes odoo/enterprise#72076

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>