[IMP] l10n_ee_reports: rework KMD and KMD INF
The tax report (KMD report) and annex to the tax report (KMD INF
report) for Estonia needed to be reworked to correct errors and
overall improve the way the KMD and KMD INF reports are presented.

With this commit, better auditing of the report lines is enabled by
adding balance_from_tags to the action_audit_cell.

This commit also rewrites the KMD INF report. It creates one report
composed of two sections with different columns, using the custom
engine to construct both parts.

Part A contains information on invoices issued, grouped by move_id
and id, to allow showing multiple lines per invoice when different
tax rates and/or special tax types are in one invoice. It only
includes invoices with taxable supply at 22%, 20%, 9% and 5%. It
excludes invoices of partners whose Tax ID is / and foreignes
partners who have a Tax ID not starting by EE.

Part B contains information on invoices received, grouped by
move_id. Each line showcases information about one invoice with
taxable supply at 22%, 20%, 9% and 5%. It relies on a Tax ID
starting with EE.

As these changes are made in stable, a warning is added to the
tax report to inform users that the module l10n_ee needs to be
upgraded and the module l10n_ee_rounding should be installed to
address rounding issues.

The reports follow the guidelines available in
https://www.emta.ee/en/business-client/taxes-and-payment/tax-returns-exchange-information/vat-return-forms-vd-and-vdp
(forms valid from 01.01.2024 onwards).

task-3997203

Part-of: odoo/enterprise#75407
Related: odoo/odoo#190180
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>