[ADD] l10n_ee_rounding: address rounding differences in tax report
This commit adds the bridge module l10n_ee_rounding, which fixes rounding
differences between Odoo and the Estonian government by adding the rounding
differences to the VAT closing entry. The difference is due to the Estonian
government computing the total sales taxes by a simple base_amount * tax_rate,
which is emulated in the tax report.

To do that, this module follows the approach implemented in l10n_fr. This
commit:
- Adds fields in res_company and in res_config_settings, setting the loss
and profit accounts for the rounding differences.
- Adds the fields to the settings view, so users can edit their values.
- Adds a default value to the fields in the Account Chart Template.
- Calls _postprocess_vat_closing_entry_results to apply the rounding
in the tax closing.

task-3997203

closes odoo/enterprise#70371

Related: odoo/odoo#180856
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>