[FIX] documents_project: rename variable created in template
Before this commit, when `hr_timesheet` and `documents_project` modules
are installed, the project variable created in the template defined
`documents_project` will erase the `project` given by the controller
route if the user goes to the portal list view of tasks via a specific
project, that is, in the route `/my/projects/<id of the project selected>`.
And so, some unexpected behavior could happen in `/my/tasks` since it
will not be expected to have a project variable set in that route.

This commit makes sure the variable created in `documents_project` will
only be used inside that template and not beeing used in other override
of that template in other modules (such as the one in `hr_timesheet`).

Steps to reproduce the issue
============================

1. Install `hr_timesheet` and `documents_project` modules
2. Go to `Project > Configuration > Projects`
3. Disable the timesheets feature on the first project.
4. Go to `/my/tasks` (portal list view of all tasks)

Actual Behavior
---------------

The cell containing the Time Spent of each task is hidden, causing a
wrong alignment for the next cells in the same row. Moreover, the cell
will be hidden even for the tasks with the timesheets feature enabled.

Expected Behavior
-----------------

The content of the cell containing the Time Spent of each task with the
timesheets feature disabled should be hidden instead of the cell itself
to avoid breaking the alignment of the next cells in the row.

task-4259996
task-4239772

closes odoo/enterprise#71984

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>