[FIX] l10n_uy_edi: bad sequence when reposting
Fixes a bug in the l10n_uy_edi module where the document number
(sequence) was not correctly reset when reposting an invoice after an
error. Previously, if an invoice was reset to draft due to an error and
reposted, Odoo would rename the document based on the previous posted
invoice. With this fix, when an invoice is reposted after being reset to
draft, the document number will be set to a generic format (*ID). This
allows the invoice to be correctly submitted to Uruware, where it will
receive the proper name, based on the next available CAE.

closes odoo/enterprise#72054

Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>