[FIX] account_reports: fix error when opening report
When there are no journals and the user tries to open the balance sheet report,
a traceback will appear.

Steps to reproduce the error:
- Install "account_reports" module without demo data
- Remove all journals
- Go Reporting > Open Balance sheet

Error: A traceback appears:
"StopIteration: null"

https://github.com/odoo/enterprise/blob/302eb1052609462fe607f77fcc934690f9abbe5a/account_reports/models/account_report.py#L379
When there are no journals "company_journals_map" will be empty,
So, it will lead to the above traceback.

sentry-5993507281

closes odoo/enterprise#72072

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>