[FIX] mrp_workorder: shop floor filter when using MO workcenter link
To reproduce:
- Create different MOs with workorders in assembly 1, in state ready
- Open shop floor
- Click on the Assembly Line 1 red link on one of the MO cards

Current behavior:
The Assembly Line 1 tab is opened, but all the workorders for this
workcenter are shown, even when they refer to different MOs.

Expected behavior:
Apply a filter on the MO when the link to the workcenter for a workorder
line on the MO card was clicked.

X-original-commit: be2e7ed9612de4544228abe4fdb8d51f457c1a74
Part-of: odoo/enterprise#73095
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Claeys Pieter (clpi) <clpi@odoo.com>