[IMP] mrp_workorder: add back "This Station" filter in shopfloor
Starting from 18.0, the default shopfloor filter "This Station" was
removed from existence. See Commit 4e6038617586b669c00a23da0ea4355f869ecc94
This filter was initially added to odoo for performance reasons [1]
and was removed since the shopfloor performance do not require it
anymore to work properly. However, this filters appeard to be usefull
for certain user in case for instance they want to filters out all the
MO's that do not have an operation which belongs to at least one of the
selected workcenter. In addition, this filter is reactive and quite
hard to manually create since it requires that the user identifies the
ids of its active workcenters.

It was therefore decided to add it back to the list of filters that can
be selected in the shopfloor in case the user would like to use it.

[1] https://github.com/odoo/enterprise/pull/64551

opw-4241699

closes odoo/enterprise#72103

Signed-off-by: William Henrotin (whe) <whe@odoo.com>