[FIX] hr_expense: allow fetching rate for ocr
Steps to reproduce:
- multi-currency
- upload an expense with a differenct currency than the company
- OCRize it

Issue:
The exchange rate is always 1:1

Cause:
We don't fetch the echange rate with OCR and the compute's cannot fetch it since we cill never satisfy the condition
https://github.com/odoo/odoo/blob/86b748dc91f7260b30c6944fd906a4d68ec1849b/addons/hr_expense/models/hr_expense.py#L251-L261

Solution:
Fetch the exchange rate directly during the process of the extracted data.

I differed the write operation at the end to avoid unwanted computed that could overwrite what we need.

opw-4151571

closes odoo/enterprise#72112

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>