[FIX] account_avatax: create taxes on parent company
Set up Avatax on a company
Create Branch and switch to it
Create an invoice with fiscal position Avatax
Compute taxes > Taxes will be created in branch company
Switch to main company
Create an invoice with fiscal position Avatax again.
Compute taxes > Error

Error will block the action
"""
Validation Error

Tax names must be unique!
- CA STATE TAX [06] (6.0000 %) in test
"""
This occurs because we first created the taxes in the branch company.
Then when retrieving taxes from avalara from the main company those
taxes are not found and will be created again but we have a constraint
that tax name must be unique per company

opw-3962212

closes odoo/enterprise#72241

X-original-commit: b5ce4f6dc40ca76cfcc45edf80f31c3d04545f1f
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>