[FIX] hr_appraisal: fixes last appraisal smart button
The issue:
When accessing the record of any employee in the employee module, we have the smart button "Last evaluation", and that button is considering the current date instead of the evaluation date of the most recent evaluation.

How to reproduce the issue:
-In the app employee on any employee, clic on the Request Appraisal button.
-Confirm the appraisal.
-Go back on the employee page.

Explanation: The behavior of this smartbutton was changed in 17.2 (https://github.com/odoo/enterprise/commit/ab01d6cbb4e2add8b6736f92c036a537bea57005).
After this fix for version 17.0, the behavior will be as follows:
-If there is an ongoing appraisal, the smartbutton display "Ongoing Appraisal" with the number of ongoing appraisals.
-If there is no ongoing appraisal, the smart button will display the date of the most recent completed appraisal, which corresponds to the date when that last appraisal was completed.

opw-4180982

closes odoo/enterprise#72253

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>