[REM] pos_restaurant_{appointment}: more options button
This commit will remove the "More Options" button from the calendar event page,
because there is a generic web option that performs the same task.

Technical
==========
As we cannot remove the views in the stable version so
this commit will just remove the more option button.
The rest of the footer, method and the inherited calendar event view of
pos_restaurant_appointment will be removed from the master

Task-4246277

closes odoo/enterprise#72176

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>