[FIX] sale_subscription: discount upsell lines according to parents
Steps to reproduce:
- Subscriptions app > New > Add 1 line with any recurring product twice
- Give both lines different discounts
- Confirm > Create Invoice > Confirm > Back to original subscription
- Upsell > Lines have the same discount

The lines should have the same discount disparity as they did on the
original SO, this is even more important when the upsell is made in
between periods as the product prices are also prorated, which makes
editing the discount manually a lot trickier. Currently both lines will
take the first parent that matches the original order line without
taking the discount into consideration.

i.e a monthly recuring order upsold halfway through the month will
already have a 50% discount (Because we only invoice half a period)
which should be cumulative with the original line's discount.

opw-4081350

closes odoo/enterprise#72406

X-original-commit: 17821cd38fc1b2e55965586b5f2cce95fc1fb519
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>