[FIX] account_budget: compute the qty_invoiced for posted moves only
Currently, the budgets rely on the `qty_invoiced` field of the Purchase
Order lines (POL). Unfortunately, this field also takes the draft bills
into account, while we don't want to account them in the budgets.

To fix this, we need to recompute the `qty_invoiced` of each POL in SQL
and filter the moves if they are not posted. So a SQL query mimicking
the compute `qty_invoiced` is added.

task-4246523

closes odoo/enterprise#74052

X-original-commit: 190eb17cdb23333a3813fafca624054067746e60
Signed-off-by: William André (wan) <wan@odoo.com>