[FIX] mrp_mps: keep domain when changing company setting
Steps:
- create 2 boms:
-- bom_A: product_A with component_A
-- bom_B: product_B with component_B
- add all 4 products to the MPS
- go to bom_A and click on the 'Schedules' smart button, there should
only be 2 records (product_A & component_A)
- open the search dropdown menu and select a 'Rows' option

Issue:
When changing the rows to be shown in the MPS (e.g. Actual Demand),
OWL calls `write` directly via the orm on res.company because those
fields are set on res.company. This is due to 'company_service.js':

```js
if (!error && model === "res.company" && UPDATE_METHODS.includes(method)) {
    if (!browser.localStorage.getItem("running_tour")) {
        action.doAction("reload_context");
    }
}
```

This cause the page to do a full refresh, which loses the domain.

Fix:
This fix makes it so that when changing the rows in the search menu,
OWL calls a specific function that will do the same job but without
the full reload, thus keeping the domain.

closes odoo/enterprise#72215

Signed-off-by: William Henrotin (whe) <whe@odoo.com>