[FIX] account_accountant: change order to check for Sign install
Steps to reproduce:
- Install Accounting
- Go to Accounting settings
- Disable "Checks" option
(A warning pops up infoming that "account_check_printing" module
will be uninstalled)
- Save

Issue:
A traceback is raised because it tries to access some fields that are
defined in "account_check_printing" module.

Cause:
After uninstallation of "account_check_printing" module, the settings
page is reloaded.
However, this reload is performed with an old version of the env
where the fields defined in the uninstalled module are still there.
It's coming from an override of "execute" method of "res.config.settings"
defined in Accounting module.
This override is checking if "sign_invoice" field is enable in order to
install Sign module.
This action is performed after the "execute" on "super()".
If the "execute" is uninstalling a module, the "env" is reset as some
models or fields may not be present anymore.
However, in the override of the "execute" method, the old "env" is
still used and non-existent fields are prefetched.

Solution:
Check to install Sign module in the override of "execute" method before
executing the one from "super()" to prevent using an expired "env".

opw-4251932

closes odoo/enterprise#72982

X-original-commit: 4890d7e088f23f65a07bf9a687485ae7fd020868
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>