[FIX] account_reports: auto-create draft tax closing moves when making a new foreign VAT fiscal position
Before this fix, when setting up a new foreign VAT fiscal position, no corresponding draft closing move was created. Because of that, if nothing was made to generate it in the meantime, when the user posted the last of the other draft closing moves, the tax lock date was set, forbidding adding new invoices for the new foreign closing, even though it wasn't posted yet. Though corner cases are still possible, we alleviate that by automatically generating the draft move when the user creates a new foreign VAT fiscal position.

closes odoo/enterprise#72548

X-original-commit: bf6edf5a14b5fdb7fd89f6fbde1510ac8c5a92fe
Related: odoo/odoo#184800
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>