[FIX] account_reports: auto-create draft tax closing moves when making a new foreign VAT fiscal position
Before this fix, when setting up a new foreign VAT fiscal position, no corresponding draft closing move was created. Because of that, if nothing was made to generate it in the meantime, when the user posted the last of the other draft closing moves, the tax lock date was set, forbidding adding new invoices for the new foreign closing, even though it wasn't posted yet. Though corner cases are still possible, we alleviate that by automatically generating the draft move when the user creates a new foreign VAT fiscal position.

closes odoo/enterprise#71857

Related: odoo/odoo#184152
Signed-off-by: William André (wan) <wan@odoo.com>