[IMP] knowledge: remove editor command "ADD_EXTERNAL_STEP"
This command's purpose was to dispatch the event at each individual external
step, but it was only used for embedded components, and `onExternalHistorySteps`
resource (which handles every batch of collaborative steps) is enough for
embedded components, and it is called less often. This commit therefore removes
`ADD_EXTERNAL_STEP` command.

task-4281356

closes odoo/enterprise#72681

Related: odoo/odoo#184929
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>