[FIX] timesheet_grid: get working hours relative to calendar tz
Steps
-----
- Install timesheet_grid
- Set monday as the first day of the week in the language settings
- Set the timezone of the user's working schedule to a more advanced
timezone, e.g. UTC+10
- Go to Timesheets and add some hours on monday
-> The under/over time is calculated relative to a lower number of
hours than the real number of working hours for that day.

Cause
-----
`get_daily_working_hours` uses UTC's start/end of day for the start and
end of the period, causing some working hours near the start/end of the
period to be missed.

opw-4163183

closes odoo/enterprise#72526

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>