[FIX] iot: display URL does not update
Due to a quirk in the JS framework, when a device
form view is opened from the IoT box form view,
it opens in a dialog box where it's `js_class`
is not run. This causes updates to Display URL
and Keyboard layout not to sync to the IoT box.

This PR works around the issue by forcing the
full device form view to open, rather than the
dialog box.

task-4141053

closes odoo/enterprise#72351

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>