[FIX] documents_project: prevent removing folders used by projects
When we remove a documents folder used by one or more projects, those projects
end up without a folder while still having the Documents feature enabled, which
should not be possible.

With this commit, there will now be an error preventing the users from doing that.

Task-4107728

Part-of: odoo/enterprise#80248
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>