[FIX] appointment_account_payment: restore the flow
ISSUE

Introduced in odoo/odoo@01b87f1230beac0568f4e3b1b76e547909506892
the payments are not always linked to account moves. As the
invoicing + appointment flow was creating the meeting on
reconciliation between the invoice account move line and the payment
account move line, using the _invoice_paid_hook method, the flow does
not work out of the box anymore.

FIX

Let us align ourselves to what we do with eCommerce. As the only
time we create a link between the invoice and the booking is
using the flow, we are creating the invoice in draft at that time,
we can consider that posting the invoice is either a manual action,
either done automatically when post processing transactions, which
would restore the previous flow, and also give more flexibility in
the backend, as manually posting the invoice will also create the
meeting.

Tests are updated.

Task-4251392

closes odoo/enterprise#72377

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>