[FIX] l10n_pk_reports: fix sign of Other Income in PL report
Steps to reproduce:
- Install l10n_pk_reports
- Switch to a Pakistani company (e.g. PK Company)
- Create a Journal entry crediting an account of "Other Income" type
(e.g. 3112003 Misc Income)
- Post the journal entry
- Go to "Accounting / Reporting / Statement Reports / Profit and Loss"
- Select "Profit and Loss (PK)" report

Issue:
The created journal entry is negative in the "Other Income" section
and its value is subtracted from "Gross Profit" section, reducing
"Profit of the Year" section, which is not correct.
On the other hand, if the normal "Profit and Loss" report is selected,
the created journal entry is positive and its value is added to "Gross
Profit" section, increasing "Net Profit" section, which is correct.

opw-4185212

closes odoo/enterprise#73253

X-original-commit: 00689949adc3673d6ff4661e56c194181c363d9d
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>