[FIX] documents: showcase relevant error while restoring
Steps to reproduce:
Create a folder in 'My Drive' e.g: "New".
Add 2 files in New folder.
Open the folder and only select those 2 files and move them to trash.
Now, go back to My drive, select that empty New folder and move it to Trash.
Go to Trash workspace.
Select one of the file and try to restore it.

Technical:
The folder whose parent has a folder_id 'False'. In the domain, we check the id
and active status for the folder_id but since there is no folder_id, that record
will not be considered as archived_top_parent_documents.

Specifications:
Show the appropriate error message when restoring from TRASH, consistent with
the error for other documents.

Task-4274176

closes odoo/enterprise#72469

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>