[FIX] documents: behavior of mail-chatter-dropzone
Steps to reproduce:
1. select a document
2. open chatter
3. drag(preview part) and drop to the others folder
4. the mail-chatter-dropzone remains open

Technical Reason:
Removing the unnecessary stop propagation on the search
panel's document drop event as it was preventing it from
being caught by the window keeping the chatter drop zone opened.

After this commit:
mail-chatter-dropzone will be closed after moved to drag documents
to the folders.

Task-4268738

closes odoo/enterprise#72443

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>