[FIX] account_reports: Aged Receivable/Payable: properly check 0-balanced elements in storno accounting
To reproduce:

1) Activate storno accounting in the settings
2) Create a new partner
3) Create a Vendor Credit Note for that partner, post it
4) Register a full payment for the credit note
5) Open the Aged Payable report
====> The partner appears in the report, even though it has an amount of 0. It shouldn't, since there is no open amount for it.

This was due to the way we checked the amount of the credit note was zero. In storno, the payable line of the refund is a negative credit instead of being a debit. Still, it's linked through the debit_move_id field to the partial reconcile. So, removing the debit from the debit or credit from the credit wasn't enough in that setup.

opw-4166891

closes odoo/enterprise#73011

X-original-commit: aeb8844be35ca7f1efe4ea1abe14559296947b18
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>