[FIX] web_studio: make placeholder_field option a subOption
This commit changes the way we display placeholder options for the CharField
when being selected by studio.

Previously, it was confusing to edit the dynamic_placeholder option, which needs
another option to work properly, and which don't make much sense in most cases.
Instead, we choose to support only the recently added placeholder_field option,
and to name it Dynamic Placeholder in the sidebar.

The feature has been tested accordingly.

task-3901821

closes odoo/enterprise#72448

Related: odoo/odoo#184657
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>