[FIX] hr_holidays_{gantt, contract_gantt}: prevent error for empty resource calendar
The error occurs because we are setting 'Working Time' to empty in 'Resources'
and then attempting to add leave from the 'Gantt' view in the Time Off section.

Steps to reproduce:
- Install 'hr_holidays_gantt' module(without demo data)
- Go to resource and from 'Administrator' remove 'Working Time'
- Now Time Off > Management > Time Off
- Go to 'Gantt' view  > Create New and Click on 'Save & Close'

Traceback: KeyError: 6

At [1], we are facing an error because 'leaves_mapping' is empty and we are
trying to access this dictionary using a key that doesn't exist.

[1]- https://github.com/odoo/enterprise/blob/d4608a66f164940fe4383507e23eed9f54bf60df/hr_holidays_gantt/models/hr_leave.py#L181

sentry-6010228921

closes odoo/enterprise#72463

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>