[PERF] documents_project: speed up lookup of project's documents
Description:
When looking for related documents on a large projects, it can
inline a large list of ids (`task_ids`) into the following
`_read_group` to retrieve the documents associated with any of the
project's tasks.
The improvement consists of using a subquery instead of a list of
ids when the list `task_ids` becomes "too large". We are using
`PREFETCH_MAX` for the "too large" threshold.

Benchmark:
The `_read_group` to fetch the related documents on 5 projects, one
of which has over 1M tasks.

|               | Before | After | Speed-up |
|---------------|--------|-------|----------|
| Timings (hot) | 2.85s  | 1.16s | 2.45x    |

closes odoo/enterprise#72603

X-original-commit: 873fa930664f0d7525b9c2b446159c34aee6a07d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>