[FIX] sign: test_ui
The shared_sign_request_tour, attempts to open some template and then click on some sign item.
However it fails when it tries to click on the sign item, because it's not found.

This happened because the tour sometimes open the wrong sign template (one with no items).
This commit aims to fix the issue by archiving other sign templates when testing this flow.

fixes runbot-98036

Part-of: odoo/enterprise#72477
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>