[FIX] account_online_synchronisation: reset fetching status
The field fetching status is used to check if we need to call the
synchronization. In case there is an error during the synchronization, the user
would be block with his connection.

The solution is to reset the fetching status when fetching the transaction or
use the new reset button on the list view of the online accounts.

task:4262788

closes odoo/enterprise#72492

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>