[FIX] whatsapp: correctly compute reply parent_id
When posting on discuss.channel, new message parent_id is constrained
as being part of the same channel. This makes sense in classic flows
as messages should belong to the same discussion thread.

However with whatsapp, parent may belong to another document. When
you post on a record using a template, replies create a new channel
that is used to discuss. In that case we want to keep the parent
as belonging to the original record.

This is necessary notably to have valid parent_id and parent record
when detecting replies in whatsapp marketing automation.

Task-4210393

closes odoo/enterprise#77811

X-original-commit: c458828e6d906811409e60038dcac9b2080109b6
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>