[FIX] sale_subscription: prevent delivered quantity products in anticipate payment
Issue: In the subscription order user portal, the anticipate payment included products with the invoice policy set to delivered quantity that were not delivered yet.

-Enable online payment and demo payment provider in the configuration.
-Create a new subscription order in the subscription module.
-Add a product with its invoice policy set to ordered quantity.
-Add another product with its invoice policy set to delivered quantity.
-Confirm the order and click on preview to access the user portal.
-Proceed with the payment through the portal.

It is now not possible to pay in advance through the user portal for recurring products with an invoice policy set to delivered quantity if they haven't been delivered yet. We still allow advance payments for partially delivered products.

opw-3775930

closes odoo/enterprise#72578

Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>