[FIX] sale_purchase_inter_company_rules: transfer order dates
When generating either a PO or SO from one company to another, currently
the commitment date of a Sale Order or the Expected Arrival of a
Purchase Order are not used to generate their counter-part in the other
company.

This means that if you set the expected arrival of the PO in Company A
to 10 days in the future, the SO generated in Company B will still try
to deliver it as soon as possible, regardless of the date set.

closes odoo/enterprise#72881

X-original-commit: 713c91230eb5140550dec0545006844d2fc7e967
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>