[FIX] spreadsheet_dashboard_hr_contract: fix contract groups
Contracts were grouped by status (New, Running, Expired)
but it's completely wrong to have a reporting on past data, because most of
those contracts would now be "expired", but they should still appear in
the "New" at the time the contract started.

With this commit, we no longer group by status but rather use the start and
end dates.

Also, the employees pivot didn't account for archived employees but they
should still be taken into account for past data.

closes odoo/enterprise#72445

Task: 4268760
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>