[IMP] appointment: prevent drag when groupby type
Prevent being able to drag an appointment type when kanban
is grouped by Type (category field) as this changes the
appointment type category and messes up the fields display.

Setting the category as readonly on the kanban view to remove
the possibility to change it on groupby via drag and drop.

Task-4282969

closes odoo/enterprise#72694

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>