[FIX] web_studio: fix approval tour operating on Apps
Before this commit, in single app mode (only studio is installed), a tour failed
because we added approvals on a button on the Form View of Apps without specifying which.
At the end of the day, a module got actually installed which triggered a lot of crashes, including
the tour itself and a few low-level cursor-related functions.

After this commit, we precisely specify in the tour on which button and function we make approvals for.
The tour succeeds and there is no crash.

runbot-error-98652

closes odoo/enterprise#72702

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>