[FIX] account_reports: fix zero-values in generic ids
When the value to store into a generic id was zero (which can typically happen when doing a non-relational groupby), None was stored into the generic id instead. On groupby lines, this lead the unfold of that line to behave weirdly and not show the proper results.

The reason was that, in Python, doing 0 == False gives True. So, when doing 0 in (False, None), we ended up with a True result.

closes odoo/enterprise#73480

X-original-commit: e280dbd8911cd7b35b2a4c44724e539a1836ef84
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>