[FIX] project_enterprise: force using UTC in test
Before this commit, the test `test_editing_task_planned_date` was written
by using the belgium timezone but this test starts failing after the end
of the summer time in belgium timezone since inside that test, the
timezone used is the one used of the current calendar set on the company
and/or the timezone set on the current user, that is the belgium one.
And so, due to the change of the hours in belgium timezone to now use
"winter time" the offset between the `datetime.now` with belgium timezone
and the one using utc timezone is no longer 2 hours but 1 hour.
Because of that the assets are totally wrong because the expected
results are fixed and do not take into account the offset could change.

This commit makes sure the test uses only utc as timezone to avoid using
offset between 2 timezones since the asserts are fixed and not dynamic
according to the offset between 2 timezones.

runbot-104873

closes odoo/enterprise#72876

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>